Major restructuring and organization of Lua API
