+++ src/application/application_context.rs
@@ -1,5 +1,5 @@
 use crate::{
-    graph::graph_compiler::CompiledProgram, lua_engine::lua_stdlib::LuaRuntime, prelude::*,
+    graph::graph_compiler::CompiledProgram, lua_engine::LuaRuntime, prelude::*,
 };
 use anyhow::Error;
 use egui_node_graph::NodeId;
